modified

+++ programming_paradigm/main-0.py
@@ -10,7 +10,7 @@ def main():
         sys.exit(1)
 
     command, *params = sys.argv[1].split(':')
-    amount = float(params[0]) if params else None
+    amount = int(params[0]) if params else None
 
     if command == "deposit" and amount is not None:
         account.deposit(amount)
@@ -27,3 +27,4 @@ def main():
 
 if __name__ == "__main__":
     main()
+
